+++ app/controllers/v1/users_delete_controller.rb
@@ -1,0 +1,52 @@
+# frozen_string_literal: true
+
+# Conovel API
+#
+# No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
+#
+# The version of the OpenAPI document: 1.0.0
+# Generated by: https://github.com/openapitools/openapi-generator.git
+#
+
+module V1
+  # UsersController
+  class UsersDeleteController < ApplicationController
+    include ErrorResponseHelper
+
+    # rubocop:disable Metrics/AbcSize
+    def delete_user_by_me
+      Rails.logger.debug("[DEBUG] カレントユーザー - @current_user_id: #{@current_user_id.to_json}")
+
+      user = User.find_by(user_id: @current_user_id)
+      if user.nil?
+        render_error_response(422, 'ユーザーが見つかりません。')
+        return
+      end
+
+      User.transaction do
+        # is_anonymous を true に設定
+        user.update!(is_anonymous: true)
+        Rails.logger.debug("[DEBUG] ユーザーの is_anonymous を true に設定しました: #{user.to_json}")
+
+        # ユーザーを削除
+        user.destroy!
+        Rails.logger.info("[INFO] ユーザーが削除されました: #{user.user_id}")
+
+        # ステータスコード 200 を返す
+        head :ok
+      rescue ActiveRecord::RecordInvalid, ActiveRecord::RecordNotDestroyed => e
+        Rails.logger.error("[ERROR] ユーザーアカウント情報の削除に失敗しました。: #{e.message}")
+        render_error_response(422, "ユーザーアカウント情報の削除に失敗しました。: #{e.message}")
+        raise ActiveRecord::Rollback
+      end
+    end
+    # rubocop:enable Metrics/AbcSize
+
+    private
+
+    # カスタムエラーメッセージを定義
+    def custom_record_invalid_message(exception)
+      "ユーザーアカウントの削除に失敗しました。: #{exception.record.errors.full_messages.join(', ')}"
+    end
+  end
+end
