+++ config/routes.rb
@@ -24,12 +24,12 @@ Rails.application.routes.draw do
     add_openapi_route 'GET', '/novels', controller_name: 'novels', action_name: 'index'
     add_openapi_route 'GET', '/sentences/{sentenceId}', controller_name: 'sentences', action_name: 'show'
     add_openapi_route 'POST', '/sentences', controller_name: 'sentences', action_name: 'create'
-    add_openapi_route 'POST', '/users/me/delete', controller_name: 'users', action_name: 'delete_user_by_me'
+    add_openapi_route 'POST', '/users/me/delete', controller_name: 'users_delete', action_name: 'delete_user_by_me'
     # add_openapi_route 'GET', '/users/{userId}/posted_novels', controller_name: 'users', action_name: 'get_novels_by_user_id'
     # add_openapi_route 'GET', '/users/{userId}', controller_name: 'users', action_name: 'show'
     add_openapi_route 'GET', '/users/me', controller_name: 'users', action_name: 'get_user_by_me'
     # add_openapi_route 'GET', '/users/me/viewedNovels', controller_name: 'users', action_name: 'get_viewed_novels_by_me'
-    add_openapi_route 'POST', '/users/me/update', controller_name: 'users', action_name: 'update_user_by_me'
+    add_openapi_route 'POST', '/users/me/update', controller_name: 'users_update', action_name: 'update_user_by_me'
   end
 
   # google認証にアクセス
